api/requirements.txt
azure-functions
langchain==0.1.0
langchain-community==0.0.10
openai==2.31.0
anthropic==0.7.1
python-dotenv==1.0.0
